software/requirements.txt
torch
transformers
Pillow
requests
numpy
scipy
datasets
